requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
python-jose[cryptography]
google-auth
requests
litellm
pydantic
python-multipart